agents/config-manager/requirements.txt
# Configuration Manager Requirements

# FastAPI framework
fastapi==0.109.0
uvicorn==0.25.0

# Pydantic for validation
pydantic==2.5.3
pydantic-settings==2.1.0

# YAML parsing
PyYAML==6.0.2

# JSON Schema validation
jsonschema==4.24.0

# Git operations
GitPython==3.1.40

# Slack notifications
slack-sdk==3.36.0

# HTTP requests
requests==2.31.0

# Python standard library (no install needed)
# - logging
# - pathlib
# - typing
# - datetime
# - os
